backend/requirements.txt
boto3==1.9.207
django-dotenv==1.4.2
django-filter==2.2.0
django-storages==1.7.2
Django==3.0.3
djangorestframework==3.11.0
Markdown==3.2
pylint-django==2.0.13
pylint==2.4.4
Pillow==7.0.0
mysqlclient==1.4.6